+++ src/Images/Pacman_closed.png
