+++ src/Images/Pacman_fullyopen_left.png
