+++ src/Images/Pacman_halfopen_down.png
